requirements.txt
# Python 3.11.14

langchain==1.0.7
langchain-core==1.0.4
langchain-community==0.4.1
langchain-chroma==1.0.0
langchain-naver==0.1.1
chromadb==1.3.4
numpy==2.3.4
tqdm==4.67.1
pymilvus==2.6.3
python-dotenv==1.2.1
neo4j==6.0.3
# Embedding & Parsing (Added for T4/Docling)
sentence-transformers==3.3.1
docling>=2.5.0
pdf2image
opencv-python-headless
langchain-huggingface
redis>=5.0.0
